Make webshell-cli pattern aware
Sometimes it is not possible to deploy a clean webshell, but to embed
the webshell into other content (e.g. a JPEG file). Previous versions of
webshell-cli did not support this, as the server response was expected
to be clean. This was now changed and webshell-cli uses patterns to
detect webshell output in server responses. All shells were already
adopted, apart from the aspx shell.